added further examples
